Almost complete

+++ Backend/src/controllers/user.controller.js
@@ -8,8 +8,9 @@ import {
 import { apiResponse } from "../utils/apiResponse.js";
 import jwt from "jsonwebtoken";
 import fs from "fs";
-import mongoose from "mongoose";
+import mongoose, { Mongoose } from "mongoose";
 import { asyncHandler } from "../utils/asyncHandler.js";
+import { pipeline } from "stream";
 
 const generateAccessAndRefreshTokens = async (userId) => {
   try {
@@ -30,66 +31,104 @@ const generateAccessAndRefreshTokens = async (userId) => {
 };
 
 const registerUser = promiseHandler(async (req, res) => {
-  // get user details from frontend
-  // validation - not empty
-  // check if user already exists: username, email
-  // check for images, check for avatar
-  // upload them to cloudinary, avatar
-  // create user object - create entry in db
-  // remove password and refresh token field from response
-  // check for user creation
-  // return res
-
-  const { fullName, email, password, username } = req.body;
+  const { fullName, email, password, username, googleUser, avatar } = req.body;
   console.log("email: ", email);
+  console.log("fullName: ", fullName);
+  console.log("password: ", password);
+  console.log("username: ", username);
 
+  // Check if all fields are provided
   if (
     [fullName, email, password, username].some((field) => field.trim() === "")
   ) {
     throw new apiError(400, "All fields are required");
   }
+
+  // Check if the user already exists with the same email or username
   const existedUser = await User.findOne({
     $or: [{ username }, { email }],
   });
+
   if (existedUser) {
-    throw new apiError(409, "User with this email or username already exists");
+    // If the user exists and is a Google user, log them in instead of throwing error
+    if (googleUser) {
+      console.log("User already exists as Google user, logging in...");
+
+      const { accessToken, refreshToken } =
+        await generateAccessAndRefreshTokens(existedUser._id);
+
+      const loggedInUser = await User.findById(existedUser._id).select(
+        "-password -refreshToken"
+      );
+
+      const options = {
+        httpOnly: true,
+        secure: true,
+      };
+
+      return res
+        .status(200)
+        .cookie("accessToken", accessToken, options)
+        .cookie("refreshToken", refreshToken, options)
+        .json(
+          new apiResponse(
+            200,
+            {
+              user: loggedInUser,
+              accessToken,
+              refreshToken,
+            },
+            "User logged in successfully"
+          )
+        );
+    } else
+      throw new apiError(
+        409,
+        "User with this email or username already exists"
+      );
   }
 
-  // console.log("req.files.avatar: ", req.files.avatar);
-
-  const avatarLocalPath = req.files?.avatar[0]?.path;
+  let avatarLocalPath;
   let coverImageLocalPath;
+
   if (
     req.files &&
     Array.isArray(req.files.coverImage) &&
     req.files.coverImage.length > 0
   ) {
-    coverImageLocalPath = req.files.coverImage[0].path;
-  }
-  if (!avatarLocalPath) {
-    throw new apiError(400, "Avatar file is required");
+    coverImageLocalPath = req.files?.coverImage[0]?.path;
   }
 
-  const avatar = await uploadOnCloudinary(avatarLocalPath);
+  if (
+    req.files &&
+    Array.isArray(req.files.avatar) &&
+    req.files.avatar.length > 0
+  ) {
+    avatarLocalPath = req.files?.avatar[0]?.path;
+  }
 
-  console.log("avatar after uploaded on cloudinary: ", avatar);
+  const Avatar = await uploadOnCloudinary(avatarLocalPath);
+  console.log("avatar after uploaded on cloudinary: ", Avatar);
 
   const coverImage = await uploadOnCloudinary(coverImageLocalPath);
 
-  if (!avatar) {
-    throw new apiError(400, "avatar file is required");
-  }
-
   const user = await User.create({
     fullName,
-    avatar: avatar.url || "",
+    avatar: Avatar?.url || avatar || "",
     email,
     password,
+    googleUser,
     coverImage: coverImage?.url || "",
     username: username.toLowerCase(),
   });
 
-  const createdUser = await User.findById(user._id).select(
+  const { accessToken, refreshToken } = await generateAccessAndRefreshTokens(
+    user._id
+  );
+  console.log("Access token: ", accessToken);
+  console.log("Refresh token: ", refreshToken);
+
+  let createdUser = await User.findById(user._id).select(
     "-password -refreshToken"
   );
 
@@ -97,23 +136,42 @@ const registerUser = promiseHandler(async (req, res) => {
     throw new apiError(500, "Error while registering user");
   }
 
+  const options = {
+    httpOnly: true,
+    secure: true,
+  };
+
+  createdUser.refreshToken = refreshToken;
+  await createdUser.save({ validateBeforeSave: false });
   req.user = createdUser;
 
   return res
     .status(201)
-    .json(new apiResponse(200, createdUser, "User registered Successfully"));
+    .cookie("accessToken", accessToken, options)
+    .cookie("refreshToken", refreshToken, options)
+    .json(
+      new apiResponse(
+        200,
+        {
+          user: createdUser,
+          accessToken,
+          refreshToken,
+        },
+        "User registered successfully"
+      )
+    );
 });
 
 const loginUser = promiseHandler(async (req, res) => {
-  const { username, password } = req.body;
+  const { username, password, googleUser } = req.body;
 
-  console.log("username: ", username, " password: ", password);
+  // console.log("username: ", username, " password: ", password);
 
   if (!username) {
     throw new apiError(404, "Username & Password are required for login");
   }
 
-  const user = await User.findOne({ username });
+  const user = await User.findOne({ username, googleUser });
 
   if (!user) {
     throw new apiError(401, "User does not exist");
@@ -128,6 +186,9 @@ const loginUser = promiseHandler(async (req, res) => {
   const { accessToken, refreshToken } = await generateAccessAndRefreshTokens(
     user._id
   );
+
+  user.refreshToken = refreshToken;
+  await user.save({ validateBeforeSave: false });
 
   const loggedInUser = await User.findById(user._id).select(
     "-password -refreshToken"
@@ -186,16 +247,19 @@ const refreshAccessToken = promiseHandler(async (req, res) => {
   if (!incomingRefreshToken) {
     throw new apiError(401, "unauthorized request");
   }
+
+  console.log("Incoming token: ", incomingRefreshToken);
+
   try {
     const decodedToken = jwt.verify(
       incomingRefreshToken,
       process.env.REFRESH_TOKEN_SECRET
     );
 
-    const user = await User.findById(decodedToken?._id);
+    let user = await User.findById(decodedToken?._id);
 
     if (!user) {
-      throw new apiError(400, "Invalid refresh token");
+      throw new apiError(403, "Invalid refresh token");
     }
 
     if (incomingRefreshToken !== user.refreshToken) {
@@ -207,21 +271,28 @@ const refreshAccessToken = promiseHandler(async (req, res) => {
       secure: true,
     };
 
-    const { accessToken, refreshToken } = generateAccessAndRefreshTokens(
-      user._id
-    );
+    // const { accessToken, refreshToken } = generateAccessAndRefreshTokens(
+    //   user._id
+    // );
 
-    return res
-      .status(200)
-      .cookie("accessToken", accessToken, options)
-      .cookie("refreshToken", refreshToken, options)
-      .json(
-        new apiResponse(
-          200,
-          { accessToken, refreshToken },
-          "Access token refreshed successfully"
+    const accessToken = user.generateAccessToken();
+
+    // user.refreshToken = refreshToken;
+    // await user.save({ validateBeforeSave: false });
+
+    return (
+      res
+        .status(200)
+        .cookie("accessToken", accessToken, options)
+        // .cookie("refreshToken", refreshToken, options)
+        .json(
+          new apiResponse(
+            200,
+            { accessToken },
+            "Access token refreshed successfully"
+          )
         )
-      );
+    );
   } catch (err) {
     throw new apiError(401, err?.message || "Invalid refresh token");
   }
@@ -249,8 +320,8 @@ const getCurrentUser = promiseHandler(async (req, res) => {
 });
 
 const updateAccountDetails = promiseHandler(async (req, res) => {
-  const { email, fullName } = req.body;
-  if (!email || !fullName) {
+  const { fullName } = req.body;
+  if (!fullName) {
     throw new apiError(404, "Missing credentials");
   }
 
@@ -258,7 +329,6 @@ const updateAccountDetails = promiseHandler(async (req, res) => {
     req.user?._id,
     {
       $set: {
-        email,
         fullName,
       },
     },
@@ -277,6 +347,8 @@ const updateAccountDetails = promiseHandler(async (req, res) => {
 const updateUserAvatar = promiseHandler(async (req, res) => {
   const avatarLocalPathTemp = req.file?.path;
 
+  // console.log("req: ", req);
+
   if (!avatarLocalPathTemp) {
     throw new apiError(404, "Avatar file missing");
   }
@@ -289,21 +361,23 @@ const updateUserAvatar = promiseHandler(async (req, res) => {
 
   console.log("url of deleting image ", req.user.avatar);
 
-  const urlParts = req.user.avatar.split("/");
-  const publicIdWithExtension = urlParts[urlParts.length - 1];
-  const publicId = publicIdWithExtension.split(".")[0];
+  if (req.user.avatar && req.user.avatar.includes("res.cloudinary.com")) {
+    const urlParts = req.user.avatar.split("/");
+    const publicIdWithExtension = urlParts[urlParts.length - 1];
+    const publicId = publicIdWithExtension.split(".")[0];
 
-  const result = await deleteImageFromCloudinary(
-    publicId,
-    function (err, result) {
-      if (err) {
-        throw new apiError(404, "Error deleting image");
+    const result = await deleteImageFromCloudinary(
+      publicId,
+      function (err, result) {
+        if (err) {
+          throw new apiError(404, "Error deleting image");
+        }
+        console.log("Deleted image successfully", result);
       }
-      console.log("Deleted image successfully", result);
-    }
-  );
+    );
 
-  console.log("result of deletion of avatar ", result);
+    console.log("result of deletion of avatar ", result);
+  }
 
   const user = await User.findByIdAndUpdate(
     req.user?._id,
@@ -334,25 +408,30 @@ const updateUserCoverImage = promiseHandler(async (req, res) => {
     throw new apiError(400, "Error while uploading cover image");
   }
 
-  const tempUser = await User.findById(req.user?._id);
+  // const tempUser = await User.findById(req.user?._id);
 
-  console.log("url of deleting image (cover image):", tempUser.coverImage);
+  // console.log("url of deleting image (cover image):", tempUser.coverImage);
 
-  const urlParts = tempUser.coverImage.split("/");
-  const publicIdWithExtension = urlParts[urlParts.length - 1];
-  const publicId = publicIdWithExtension.split(".")[0];
+  if (
+    req.user.coverImage &&
+    req.user.coverImage.includes("res.cloudinary.com")
+  ) {
+    const urlParts = tempUser.coverImage.split("/");
+    const publicIdWithExtension = urlParts[urlParts.length - 1];
+    const publicId = publicIdWithExtension.split(".")[0];
 
-  const result = await deleteImageFromCloudinary(
-    publicId,
-    function (err, result) {
-      if (err) {
-        throw new apiError(404, "Error deleting image");
+    const result = await deleteImageFromCloudinary(
+      publicId,
+      function (err, result) {
+        if (err) {
+          throw new apiError(404, "Error deleting image");
+        }
+        console.log("Deleted image successfully", result);
       }
-      console.log("Deleted image successfully", result);
-    }
-  );
+    );
 
-  console.log("result of deletion of avatar ", result);
+    console.log("result of deletion of avatar ", result);
+  }
 
   const user = await User.findByIdAndUpdate(
     req.user?._id,
@@ -499,6 +578,220 @@ const getWatchHistory = promiseHandler(async (req, res) => {
     );
 });
 
+const getUserDetails = asyncHandler(async (req, res) => {
+  const { userId } = req.params;
+
+  if (!userId) {
+    throw new apiError(400, "User id is necessary for getting user details");
+  }
+
+  const userDetails = await User.aggregate([
+    {
+      $match: {
+        _id: new mongoose.Types.ObjectId(userId),
+      },
+    },
+    {
+      $lookup: {
+        from: "videos",
+        localField: "_id",
+        foreignField: "owner",
+        as: "videos",
+        pipeline: [
+          {
+            $project: {
+              _id: 1,
+              videoFile: 1,
+              thumbnail: 1,
+              duration: 1,
+              views: 1,
+              isPublished: 1,
+              createdAt: 1,
+            },
+          },
+          {
+            $lookup: {
+              from: "comments",
+              localField: "_id",
+              foreignField: "video",
+              as: "comments",
+              pipeline: [
+                {
+                  $project: {
+                    _id: 1,
+                    content: 1,
+                    owner: 1,
+                  },
+                },
+                {
+                  $lookup: {
+                    from: "likes",
+                    localField: "_id",
+                    foreignField: "comment",
+                    as: "likes",
+                    pipeline: [
+                      {
+                        $project: {
+                          _id: 1,
+                          likedBy: 1,
+                        },
+                      },
+                    ],
+                  },
+                },
+              ],
+            },
+          },
+          {
+            $lookup: {
+              from: "likes",
+              localField: "_id",
+              foreignField: "video",
+              as: "likes",
+              pipeline: [
+                {
+                  $project: {
+                    _id: 1,
+                    likedBy: 1,
+                  },
+                },
+              ],
+            },
+          },
+        ],
+      },
+    },
+    {
+      $lookup: {
+        from: "subscriptions",
+        localField: "_id",
+        foreignField: "channel",
+        as: "subscribers",
+        pipeline: [
+          {
+            $project: {
+              subscriber: 1,
+            },
+          },
+        ],
+      },
+    },
+    {
+      $lookup: {
+        from: "playlists",
+        localField: "_id",
+        foreignField: "owner",
+        as: "playlists",
+        pipeline: [
+          {
+            $lookup: {
+              from: "videos",
+              localField: "videos",
+              foreignField: "_id",
+              as: "videos",
+            },
+          },
+          {
+            $addFields: {
+              videosLength: {
+                $size: "$videos",
+              },
+              videos: {
+                $first: "$videos",
+              },
+            },
+          },
+        ],
+      },
+    },
+    {
+      $addFields: {
+        subscribers: {
+          $first: "$subscribers",
+        },
+      },
+    },
+  ]);
+
+  if (!userDetails) {
+    throw new apiError(400, "No such user exist");
+  }
+
+  return res
+    .status(201)
+    .json(
+      new apiResponse(200, userDetails[0], "User details fetched successfully")
+    );
+});
+
+const addWatchHistory = asyncHandler(async (req, res) => {
+  const { videoId } = req.params;
+
+  if (!videoId) {
+    throw new apiError(400, "Video id is required for adding watch history");
+  }
+
+  let user = await User.findById(req.user?._id);
+
+  if (!user) {
+    throw new apiError(400, "User not found");
+  }
+
+  if (!user.watchHistory?.includes(videoId)) {
+    user.watchHistory.push(videoId);
+    await user.save({ validateBeforeSave: false });
+  }
+
+  console.log("After added watch history: ", user);
+
+  return res
+    .status(201)
+    .json(new apiResponse(200, user, "Video added to watch history"));
+});
+
+const removeVideoHistory = asyncHandler(async (req, res) => {
+  const { videoId } = req.params;
+
+  if (!videoId) {
+    throw new apiError(
+      400,
+      "Video id is required for remove video from watch history"
+    );
+  }
+
+  let user = await User.findById(req.user?._id);
+
+  if (!user) {
+    throw new apiError(400, "User not found");
+  }
+
+  if (user.watchHistory?.includes(videoId)) {
+    user.watchHistory = user.watchHistory.filter(
+      (watch) => watch.toString() !== videoId.toString()
+    );
+    await user.save({ validateBeforeSave: false });
+  }
+
+  return res
+    .status(200)
+    .json(new apiResponse(201, user, "Removed video from history"));
+});
+
+const clearWatchHistory = asyncHandler(async (req, res) => {
+  let user = await User.findById(req.user?._id);
+
+  if (!user) {
+    throw new apiError(400, "User not found");
+  }
+
+  user.watchHistory = [];
+  await user.save({ validateBeforeSave: false });
+
+  return res
+    .status(201)
+    .json(new apiResponse(200, user, "Clear watch history"));
+});
+
 export {
   registerUser,
   loginUser,
@@ -511,4 +804,8 @@ export {
   updateUserCoverImage,
   getUserChannelProfile,
   getWatchHistory,
+  getUserDetails,
+  addWatchHistory,
+  removeVideoHistory,
+  clearWatchHistory,
 };
